# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Taif Children's Hospital Research Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
...
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List documents
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentList"
  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get document by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete document (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
...
  /articles:
    get:
      operationId: listArticles
      tags: [articles]
      summary: List articles
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of articles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleList"
    post:
      operationId: createArticle
      tags: [articles]
      summary: Create article
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateArticleBody"
      responses:
        "201":
          description: Created article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
  /articles/{id}:
    get:
      operationId: getArticle
      tags: [articles]
      summary: Get article by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
    patch:
      operationId: updateArticle
      tags: [articles]
      summary: Update article
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateArticleBody"
      responses:
        "200":
          description: Updated article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
    delete:
      operationId: deleteArticle
      tags: [articles]
      summary: Delete article
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
  /calendar/events:
    get:
      operationId: listEvents
      tags: [calendar]
      summary: List calendar events
      parameters:
        - name: start
          in: query
          schema:
            type: string
        - name: end
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalendarEvent"
    post:
      operationId: createEvent
      tags: [calendar]
      summary: Create calendar event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventBody"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarEvent"
  /calendar/events/{id}:
    patch:
      operationId: updateEvent
      tags: [calendar]
      summary: Update calendar event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEventBody"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarEvent"
    delete:
      operationId: deleteEvent
      tags: [calendar]
      summary: Delete calendar event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
...
components:
  schemas:
...
    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        full_name:
          type: ["string", "null"]
        full_name_ar:
          type: ["string", "null"]
        role:
          type: string
        department:
          type: ["string", "null"]
        avatar_url:
          type: ["string", "null"]
        is_active:
          type: boolean
        created_at:
          type: string
        last_sign_in_at:
          type: ["string", "null"]
      required: [id, email, role, is_active, created_at]
...
    UpdateUserBody:
      type: object
      properties:
        full_name:
          type: ["string", "null"]
        full_name_ar:
          type: ["string", "null"]
        role:
          type: ["string", "null"]
        department:
          type: ["string", "null"]
        is_active:
          type: ["boolean", "null"]
...
    Document:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        title_ar:
          type: ["string", "null"]
        file_name:
          type: string
        file_type:
          type: string
        file_size:
          type: integer
        file_url:
          type: string
        uploaded_by:
          type: string
        uploader_name:
          type: ["string", "null"]
        created_at:
          type: string
        description:
          type: ["string", "null"]
      required: [id, title, file_name, file_type, file_size, file_url, uploaded_by, created_at]
    DocumentList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Document"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [items, total, page, limit]
...
    Article:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        title_ar:
          type: ["string", "null"]
        content:
          type: string
        content_ar:
          type: ["string", "null"]
        excerpt:
          type: ["string", "null"]
        excerpt_ar:
          type: ["string", "null"]
        cover_image_url:
          type: ["string", "null"]
        author_id:
          type: string
        author_name:
          type: ["string", "null"]
        is_published:
          type: boolean
        published_at:
          type: ["string", "null"]
        created_at:
          type: string
        updated_at:
          type: string
      required: [id, title, content, author_id, is_published, created_at, updated_at]
    ArticleList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Article"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [items, total, page, limit]
    CreateArticleBody:
      type: object
      properties:
        title:
          type: string
        title_ar:
          type: ["string", "null"]
        content:
          type: string
        content_ar:
          type: ["string", "null"]
        excerpt:
          type: ["string", "null"]
        excerpt_ar:
          type: ["string", "null"]
        cover_image_url:
          type: ["string", "null"]
        is_published:
          type: boolean
      required: [title, content, is_published]
    UpdateArticleBody:
      type: object
      properties:
        title:
          type: ["string", "null"]
        title_ar:
          type: ["string", "null"]
        content:
          type: ["string", "null"]
        content_ar:
          type: ["string", "null"]
        excerpt:
          type: ["string", "null"]
        excerpt_ar:
          type: ["string", "null"]
        cover_image_url:
          type: ["string", "null"]
        is_published:
          type: ["boolean", "null"]
    CalendarEvent:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        title_ar:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        description_ar:
          type: ["string", "null"]
        event_type:
          type: string
        start_time:
          type: string
        end_time:
          type: ["string", "null"]
        all_day:
          type: boolean
        location:
          type: ["string", "null"]
        created_by:
          type: string
        creator_name:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        created_at:
          type: string
      required: [id, title, event_type, start_time, all_day, created_by, created_at]
    CreateEventBody:
      type: object
      properties:
        title:
          type: string
        title_ar:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        description_ar:
          type: ["string", "null"]
        event_type:
          type: string
        start_time:
          type: string
        end_time:
          type: ["string", "null"]
        all_day:
          type: boolean
        location:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
      required: [title, event_type, start_time, all_day]
    UpdateEventBody:
      type: object
      properties:
        title:
          type: ["string", "null"]
        title_ar:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        description_ar:
          type: ["string", "null"]
        event_type:
          type: ["string", "null"]
        start_time:
          type: ["string", "null"]
        end_time:
          type: ["string", "null"]
        all_day:
          type: ["boolean", "null"]
        location:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
...